refactor(lcode_backend): harmonize the metadata of LCODE particle data (momentum) to standard names and units, update docs and tests
The third momentum component in LCODE particle data corresponds to either $p_z$ in Cartesian geometry or the angular momentum $L$ in axisymmetric/cylindrical geometry. This is now taken into account via the boolean parameter `axisym` (`True` by default). In cylindrical geometry the third momentum component is renamed and a new `'p3'` variable is added to the dataset, corresponding to $p_\theta = L / r$.

In addition, all momenta in LCODE particle data are normalised to $m_e \ c$. The units are now converted to $m_\mathrm{sp} \ c$, using the charge-to-mass ratio in the data and the new argument `abs_q` (absolute value of the normalised bunch particle charge).