fix(part_mixin): error when trying to save the output of `get_slice_emittance` or `get_energy_spectrum`
These particle methods use a special binning and grouping function (`flox.xarray.xarray_reduce`) for increased efficiency. However, this function creates a coordinate for the binned variable with the data type `pandas.Interval`, which makes it impossible to save the data with `ds.ozzy.save`. The binned coordinate is now converted into a normal `numpy` array, using the midpoints of each interval element.