feat(part_mixin): add function to get the weighted median of some variable in a particle dataset
Particle data from PIC codes is often weighted (often in the form of an individual macroparticle "charge"), which means that the calculation of a median value is not as straightforward as for other discrete and unweighted data. This method ensures that a median value can be calculated correctly, taking the macroparticle weights into account.

This function can be called as a particle method, e.g. `part_ds.ozzy.get_weighted_mean`, where `part_ds` is an ozzy particle Dataset object (i.e., with attribute `pic_data_type = "part"`).