refactor: add argument to `xarray.concat` and `xarray.merge` calls due to xarray `FutureWarning`
Add the argument `join="outer"` explicitly when call `xarray.concat` to avoid getting an error in the future, when xarray changes the default value of `join` from `"outer"` to `"exact"`.
See warning:
```
FutureWarning: In a future version of xarray the default value for join will change from join='outer' to join='exact'.
```